+++ scripts/libtime.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+# SPDX-License-Identifier: (LGPL-2.1 OR BSD-2-Clause)
+# Copyright (C) 2026 Rong Tao. All rights reserved.
+
+get_nsecs() {
+	date +%s%9N
+}