+++ scripts/libtime_examples.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+set -e
+
+. libtime.sh
+. liblog.sh
+
+get_nsecs